Notebook as it was after teaching on 20151218
This was the notebook I used to teach a session to two people at
Citadel, on Dec 18th, 2015. I created cells and changed a little bit of
the content on the fly, and the notebook might not execute cleanly from
beginning to end. For that reason, I'm checking in the executed
notebook, with all the outputs.